requirements.txt
torch>=2.3.0
transformers>=4.45.0
accelerate>=0.34.0
datasets>=2.20.0
evaluate>=0.4.2
bitsandbytes>=0.43.0
sentencepiece>=0.2.0
protobuf>=4.25.0
tqdm>=4.66.0
pyyaml>=6.0
numpy>=1.26.0
pandas>=2.2.0
matplotlib>=3.9.0
seaborn>=0.13.0
huggingface-hub>=0.25.0
